Rename external tags components for clarity
Issue #11
