+++ kolibri_dynamic_collections_plugin/assets/src/components/TagsListEditor.vue
